+++ src/main/java/nextstep/ladder/domain/line/LadderHeight.java
@@ -5,7 +5,6 @@ import java.util.stream.IntStream;
 public class LadderHeight {
 
     private static final int START = 0;
-    private static final String INVALID_LADDER_HEIGHT = "사다리의 높이는 1이상이여야 합니다.";
 
     private final int height;
 
@@ -20,7 +19,7 @@ public class LadderHeight {
 
     private void validate(int height) {
         if (height < 1) {
-            throw new IllegalArgumentException(INVALID_LADDER_HEIGHT);
+            throw new IllegalArgumentException("사다리의 높이는 1이상이여야 합니다.");
         }
     }
 }
