+++ src/main/java/nextstep/ladder/view/ResultView.java
@@ -4,9 +4,13 @@ import nextstep.ladder.domain.line.Ladder;
 import nextstep.ladder.domain.line.LadderHeight;
 import nextstep.ladder.domain.user.Users;
 
+import java.util.stream.IntStream;
+
 
 public class ResultView {
 
+    private static final int START = 0;
+    private static final int WIDTH_OUT_OF_RANGE_BLOCK_NUMBER = 1;
     private static final String RESULT = "\n실행결과";
     private static final String HEIGHT_SEPARATOR = "|";
     private static final String LINE_SEPARATOR  = "-----";
@@ -25,7 +29,7 @@ public class ResultView {
     }
 
     private void printHeight(Ladder ladder, int height, Users users) {
-        users.toStream().forEach(width -> {
+        IntStream.range(START, users.userCount() - WIDTH_OUT_OF_RANGE_BLOCK_NUMBER).forEach(width -> {
             System.out.print(HEIGHT_SEPARATOR);
             printLine(ladder, height, width);
         });
