+++ src/test/java/nextstep/ladder/domain/line/LadderHeightTest.java
@@ -7,7 +7,7 @@ import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException
 class LadderHeightTest {
 
     @Test
-    public void LadderHeight_생성테스트() {
+    public void Invalid_사다리높이() {
         assertThatIllegalArgumentException().isThrownBy(() -> new LadderHeight(0));
     }
 }
