+++ src/test/java/nextstep/ladder/domain/line/LineTest.java
@@ -1,6 +1,5 @@
 package nextstep.ladder.domain.line;
 
-import nextstep.ladder.domain.line.generator.TestLineGenerator;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
